% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user.R
\name{ask_1_2}
\alias{ask_1_2}
\title{Fonction pour interagir avec l'utilisateur}
\usage{
ask_1_2(question, unattended_option = NULL)
}
\arguments{
\item{question}{La question à poser}
}
\value{
La réponse de l'utilisateur
}
\description{
Fonctions pour poser des questions à l'utilisateur à laquelle
il peut répondre par 1 ou 2
}
