% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_user.R
\name{ask_yes_no}
\alias{ask_yes_no}
\title{Fonction pour interagir avec l'utilisateur}
\usage{
ask_yes_no(question, unattended_option = NULL)
}
\arguments{
\item{question}{La question \\u00e0 poser}

\item{unattended_option}{Option pour mode non-interactif ("oui"/"non")}
}
\value{
La r\\u00e9ponse de l'utilisateur
}
\description{
Fonctions pour poser des questions \\u00e0 l'utilisateur \\u00e0 laquelle
il peut r\\u00e9pondre par oui ou non
}
