% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_coerce_timestamp_tz}
\alias{athena_coerce_timestamp_tz}
\title{Convertir une colonne Athena \verb{timestamp with time zone}}
\usage{
athena_coerce_timestamp_tz(x, default_timezone = "UTC")
}
\arguments{
\item{x}{Un vecteur de chaînes "yyyy-mm-dd HH:MM:SS.sss \if{html}{\out{<zone>}}"}

\item{default_timezone}{Fuseau de repli si la zone embarquée est invalide}
}
\value{
Un vecteur \code{POSIXct}
}
\description{
Convertir une colonne Athena \verb{timestamp with time zone}
}
\keyword{internal}
