% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_dbi}
\alias{athena_dbi}
\alias{dbConnect,EllipseAthenaDriver-method}
\alias{dbDisconnect,EllipseAthenaConnection-method}
\alias{dbIsValid,EllipseAthenaConnection-method}
\alias{dbGetInfo,EllipseAthenaConnection-method}
\alias{dbSendQuery,EllipseAthenaConnection-method}
\alias{dbFetch,EllipseAthenaResult-method}
\alias{dbHasCompleted,EllipseAthenaResult-method}
\alias{dbClearResult,EllipseAthenaResult-method}
\alias{dbColumnInfo,EllipseAthenaResult-method}
\alias{dbListTables,EllipseAthenaConnection-method}
\alias{dbExistsTable,EllipseAthenaConnection,character-method}
\title{Athena DBI driver/connection/result classes}
\usage{
\S4method{dbConnect}{EllipseAthenaDriver}(
  drv,
  aws_access_key_id,
  aws_secret_access_key,
  profile_name,
  schema_name,
  s3_staging_dir,
  work_group = NULL,
  ...
)

\S4method{dbDisconnect}{EllipseAthenaConnection}(conn, ...)

\S4method{dbIsValid}{EllipseAthenaConnection}(dbObj, ...)

\S4method{dbGetInfo}{EllipseAthenaConnection}(dbObj, ...)

\S4method{dbSendQuery}{EllipseAthenaConnection}(conn, statement, ...)

\S4method{dbFetch}{EllipseAthenaResult}(res, n = -1, ...)

\S4method{dbHasCompleted}{EllipseAthenaResult}(res, ...)

\S4method{dbClearResult}{EllipseAthenaResult}(res, ...)

\S4method{dbColumnInfo}{EllipseAthenaResult}(res, ...)

\S4method{dbListTables}{EllipseAthenaConnection}(conn, ...)

\S4method{dbExistsTable}{EllipseAthenaConnection,character}(conn, name, ...)
}
\description{
\code{tube} implémente son propre pilote DBI pour AWS Athena, basé directement
sur \code{paws.analytics::athena()}. Ce pilote remplace le package \code{noctua}
(retiré de CRAN), sans ajouter de nouvelle dépendance puisque
\code{paws.analytics} est déjà utilisé ailleurs dans le package (voir R/glue.R).
}
\details{
Les classes héritent des classes S4 de \code{DBI} (\code{DBIDriver}, \code{DBIConnection},
\code{DBIResult}), ce qui permet à \code{dplyr}/\code{dbplyr} de générer et exécuter des
requêtes paresseuses (\code{dplyr::tbl()}) sans traduction SQL personnalisée:
\code{dbplyr} retombe automatiquement sur ses méthodes génériques pour
\code{DBIConnection}, qui conviennent au dialecte SQL d'Athena (Presto/Trino).
}
