% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_ellipse_driver}
\alias{athena_ellipse_driver}
\title{Construire le pilote DBI Athena interne de tube}
\usage{
athena_ellipse_driver()
}
\value{
Un objet \code{EllipseAthenaDriver}
}
\description{
Construire le pilote DBI Athena interne de tube
}
\keyword{internal}
