% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_execute_query}
\alias{athena_execute_query}
\title{Exécuter une requête Athena et attendre sa complétion}
\usage{
athena_execute_query(conn, statement, poll_interval_sec = 1, timeout_sec = 300)
}
\arguments{
\item{conn}{Un objet \code{EllipseAthenaConnection}}

\item{statement}{La requête SQL à exécuter}

\item{poll_interval_sec}{Délai (en secondes) entre deux vérifications d'état}

\item{timeout_sec}{Délai maximal (en secondes) à attendre avant d'abandonner}
}
\value{
L'identifiant de l'exécution de requête (\code{QueryExecutionId})
}
\description{
Exécuter une requête Athena et attendre sa complétion
}
\keyword{internal}
