% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_fetch_results}
\alias{athena_fetch_results}
\title{Récupérer, en paginant, tous les résultats d'une requête Athena complétée}
\usage{
athena_fetch_results(conn, query_execution_id, statement)
}
\arguments{
\item{conn}{Un objet \code{EllipseAthenaConnection}}

\item{query_execution_id}{L'identifiant retourné par \code{athena_execute_query()}}

\item{statement}{La requête SQL à exécuter}
}
\value{
Un \code{tibble} contenant les résultats de la requête
}
\description{
Récupérer, en paginant, tous les résultats d'une requête Athena complétée
}
\keyword{internal}
