% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.R
\name{aws_session}
\alias{aws_session}
\title{Create an AWS session object}
\usage{
aws_session(id = NULL, key = NULL)
}
\arguments{
\item{id}{An AWS access key id}

\item{key}{An AWS secret access key}
}
\value{
An AWS session object
}
\description{
Create an AWS session object
}
