% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{categorize_directory_items}
\alias{categorize_directory_items}
\title{Categorize directory items into supported files, folders, and other files}
\usage{
categorize_directory_items(current_dir, items)
}
\description{
Categorize directory items into supported files, folders, and other files
}
\keyword{internal}
