% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_database}
\alias{check_database}
\title{Check the database parameter provided to a function}
\usage{
check_database(database)
}
\arguments{
\item{database}{The database to check}
}
\value{
TRUE if the database parameter is valid, FALSE otherwise
}
\description{
Check if the database parameter is valid : either datawarehouse, datamarts, or datalake
}
