% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_env}
\alias{check_env}
\title{Check the env parameter provided to a function}
\usage{
check_env(env)
}
\arguments{
\item{env}{The environment to check}
}
\value{
TRUE if the env parameter is valid, FALSE otherwise
}
\description{
Check if the env parameter is valid: either DEV or PROD
}
