% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_file_versioning_before_ingest}
\alias{check_file_versioning_before_ingest}
\title{Check the parameters provided to the ellipse_ingest function}
\usage{
check_file_versioning_before_ingest(file_batch, file_version)
}
\arguments{
\item{file_batch}{The batch value of the file to ingest that is going to
be transformed into a column in the datawarehouse table}

\item{file_version}{The version of the file to ingest that is going to
be transformed into a column in the datawarehouse table}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before ingesting the data
}
