% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{check_landing_zone}
\alias{check_landing_zone}
\title{Checks is the files in the landing zone have been processed and tells the user if he can
upload more}
\usage{
check_landing_zone(creds, pipeline_name)
}
\arguments{
\item{creds}{An AWS session object with your credentials and the aws ressources required}

\item{pipeline_name}{The name of the pipeline (i.e.: the name of the first folder in the path within the landding zone bucket)}
}
\value{
the number of unprocessed files in the landing zone
}
\description{
Checks is the files in the landing zone have been processed and tells the user if he can
upload more
}
\examples{
\dontrun{
 r <- check_landing_zone(aws_session(), "my_pipeline")
print(r)
}
}
