% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_params_before_describe}
\alias{check_params_before_describe}
\title{Check the parameters provided to the ellipse_describe function}
\usage{
check_params_before_describe(
  env,
  schema,
  table,
  new_table_tags,
  new_table_description
)
}
\arguments{
\item{env}{The environment to describe the data from}

\item{schema}{The datamart to describe the data from}

\item{table}{The table to describe the data from}

\item{new_table_tags}{The new table tags to apply to the table}

\item{new_table_description}{The new table description to apply to the table}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before executing the function
}
