% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_params_before_publish}
\alias{check_params_before_publish}
\title{Check the parameters provided to the ellipse_publish function}
\usage{
check_params_before_publish(
  env,
  dataframe,
  datamart,
  table,
  data_tag,
  table_tags,
  table_description
)
}
\arguments{
\item{env}{The environment to publish the data to}

\item{dataframe}{The dataframe to publish}

\item{datamart}{The datamart to publish the data to}

\item{table}{The table to publish the data to}

\item{data_tag}{The data tag for the published data}

\item{table_tags}{The table tags to associate with the published table}

\item{table_description}{The description to associate with the published table}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before publishing the data
}
