% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_params_before_refresh}
\alias{check_params_before_refresh}
\title{Check the parameters provided to the ellipse_process function}
\usage{
check_params_before_refresh(con, schema, table)
}
\arguments{
\item{con}{The connection to the database}

\item{schema}{The datamart to refresh the table from}

\item{table}{The table to run the glue job against}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before executing the function
}
