% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_params_before_unpublish}
\alias{check_params_before_unpublish}
\title{Check the parameters provided to the ellipse_unpublish function}
\usage{
check_params_before_unpublish(env, datamart, table)
}
\arguments{
\item{env}{The environment to unpublish the data from}

\item{datamart}{The datamart to unpublish the data from}

\item{table}{The table to unpublish the data from}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before unpublishing the data
}
