% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_pipeline_before_ingest}
\alias{check_pipeline_before_ingest}
\title{Check the parameters provided to the ellipse_ingest function}
\usage{
check_pipeline_before_ingest(
  pipeline,
  landing_zone_partitions,
  file_batch,
  file_version
)
}
\arguments{
\item{pipeline}{The pipeline to check}

\item{landing_zone_partitions}{The list of partitions in the landing zone bucket which
corresponds to the pipelines names that have been implemented in the platform}

\item{file_batch}{The batch value of the file to ingest that is going to
be transformed into a column in the datawarehouse table}

\item{file_version}{The version of the file to ingest that is going to
be transformed into a column in the datawarehouse table}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before ingesting the data
}
