% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classify_base}
\alias{classify_base}
\title{Base classification algorithm}
\usage{
classify_base(xtrain, ytrain, xnew, method, ...)
}
\arguments{
\item{xtrain}{A matrix with columns representing features and rows representing observations.}

\item{ytrain}{A binary vector specifying the corresponding labels of the observations.}

\item{xnew}{A matrix specifying new observations whose class labels need to be predicted.}

\item{method}{A character specifying the base classification method to be used. Should be one of
"LR" (logistic regression), "penLR" (penalized logistic regression)
"RF" (random forest), "GB" (gradient boosting), and "NB" (naive Bayes).}

\item{...}{Additional arguments parsed to the algorithms.}
}
\value{
A vector of classification scores for \code{xnew}.
}
\description{
Base classification algorithm
}
