% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{collect_all_metadata_interactive}
\alias{collect_all_metadata_interactive}
\title{Collect ALL metadata including required system fields}
\usage{
collect_all_metadata_interactive()
}
\description{
Collect ALL metadata including required system fields
}
\keyword{internal}
