% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{collect_custom_metadata_interactive}
\alias{collect_custom_metadata_interactive}
\title{Collect custom metadata interactively}
\usage{
collect_custom_metadata_interactive()
}
\description{
Collect custom metadata interactively
}
\keyword{internal}
