% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{collect_optional_custom_metadata}
\alias{collect_optional_custom_metadata}
\title{Collect optional custom metadata fields}
\usage{
collect_optional_custom_metadata(metadata)
}
\description{
Collect optional custom metadata fields
}
\keyword{internal}
