% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{confirm_directory_selection}
\alias{confirm_directory_selection}
\title{Confirm directory selection and show preview}
\usage{
confirm_directory_selection(dir_path)
}
\description{
Confirm directory selection and show preview
}
\keyword{internal}
