% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{confirm_simple_directory_selection}
\alias{confirm_simple_directory_selection}
\title{Confirm directory selection with simple preview}
\usage{
confirm_simple_directory_selection(dir_path)
}
\description{
Confirm directory selection with simple preview
}
\keyword{internal}
