% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{dbplyr_edition.EllipseAthenaConnection}
\alias{dbplyr_edition.EllipseAthenaConnection}
\title{Utiliser l'interface moderne (edition 2) de dbplyr pour dplyr::tbl()}
\usage{
\method{dbplyr_edition}{EllipseAthenaConnection}(con)
}
\arguments{
\item{con}{Un objet \code{EllipseAthenaConnection}}
}
\value{
\code{2L}
}
\description{
Utiliser l'interface moderne (edition 2) de dbplyr pour dplyr::tbl()
}
