% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{delete_glue_table}
\alias{delete_glue_table}
\title{Delete glue table}
\usage{
delete_glue_table(credentials, database_name, table_name)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{database_name}{The name of the database to delete the table from}

\item{table_name}{The name of the table to delete}
}
\value{
A boolean indicating wether or not the table was deleted
successfully
}
\description{
Delete glue table
}
