% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{delete_s3_folder}
\alias{delete_s3_folder}
\title{Delete s3 folder}
\usage{
delete_s3_folder(credentials, bucket, prefix)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{bucket}{The bucket to delete the folder from}

\item{prefix}{The prefix (or partition) to delete}
}
\value{
TRUE if the folder was deleted successfully
}
\description{
Delete s3 folder
}
