% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_discover.R
\name{display_categorized_datasets}
\alias{display_categorized_datasets}
\title{Display categorized datasets in separate sections}
\usage{
display_categorized_datasets(categorized_datasets)
}
\arguments{
\item{categorized_datasets}{List from categorize_datasets_by_content}
}
\description{
Display categorized datasets in separate sections
}
\keyword{internal}
