% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{display_html_file}
\alias{display_html_file}
\title{Display an HTML file by opening it in a web browser}
\usage{
display_html_file(filepath)
}
\arguments{
\item{filepath}{Path to HTML file to display}
}
\value{
Invisibly returns the filepath
}
\description{
Opens an HTML file in the default web browser with multiple fallback options for
Linux/Mac/Windows. Tries multiple browser opening methods in order of preference:
\itemize{
\item Linux: xdg-open, firefox, chrome/chromium
\item Mac: open command
\item Windows: start command
}
}
\details{
Temp files should persist as browser needs time to load them.
}
\seealso{
\code{\link{display_image_file}} for image display
}
\keyword{internal}
