% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{display_image_file}
\alias{display_image_file}
\title{Display an image file using R's built-in viewer capabilities}
\usage{
display_image_file(filepath)
}
\arguments{
\item{filepath}{Path to image file}
}
\description{
Display an image file using R's built-in viewer capabilities
}
\keyword{internal}
