% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_query.R
\name{display_image_from_s3}
\alias{display_image_from_s3}
\title{Display an image from S3 using R's built-in viewer}
\usage{
display_image_from_s3(file_info, credentials)
}
\arguments{
\item{file_info}{Single row from files_metadata}

\item{credentials}{AWS credentials}
}
\description{
Display an image from S3 using R's built-in viewer
}
\keyword{internal}
