% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{display_simple_file_menu}
\alias{display_simple_file_menu}
\title{Display simple file menu with literal navigation and numbered choices}
\usage{
display_simple_file_menu(items)
}
\description{
Display simple file menu with literal navigation and numbered choices
}
\keyword{internal}
