% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_query.R
\name{download_and_aggregate_files}
\alias{download_and_aggregate_files}
\title{Download files from S3 and aggregate into single dataframe}
\usage{
download_and_aggregate_files(files_metadata, credentials)
}
\arguments{
\item{files_metadata}{Tibble with file metadata}

\item{credentials}{AWS credentials from get_aws_credentials()}
}
\description{
Download files from S3 and aggregate into single dataframe
}
\keyword{internal}
