% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{download_s3_file_to_temp}
\alias{download_s3_file_to_temp}
\title{Download S3 file to temporary location}
\usage{
download_s3_file_to_temp(s3_path, credentials)
}
\arguments{
\item{s3_path}{S3 path (s3://bucket/key format)}

\item{credentials}{AWS credentials from get_aws_credentials()}
}
\description{
Download S3 file to temporary location
}
\keyword{internal}
