% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_connect}
\alias{ellipse_connect}
\title{Connexion à la plateforme de données ellipse sur AWS}
\usage{
ellipse_connect()
}
\value{
Un object de connexion \code{DBI}.
}
\description{
Cette fonction utilise les clés d'accès AWS configurées dans le fichier
\code{.Renviron} pour se connecter à la plateforme de données.
}
