% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_disconnect}
\alias{ellipse_disconnect}
\title{Se déconnecter de la plateforme de données ellipse}
\usage{
ellipse_disconnect(con = NULL)
}
\arguments{
\item{con}{L'objet de connexion à fermer}
}
\value{
TRUE if the connexion was closed or FALSE if no connexion existed
}
\description{
Se déconnecter de la plateforme de données ellipse
}
