% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_process}
\alias{ellipse_process}
\title{Traiter les données en attente d'être insérées dans une table}
\usage{
ellipse_process(con, table)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{table}{Le nom de la table qui doit traitée}
}
\value{
TRUE si les données en attente d'être insérées ont bien été traitées.
FALSE sinon.
}
\description{
Traiter les données en attente d'être insérées dans une table
}
