% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{ellipse_push_datalake_mode}
\alias{ellipse_push_datalake_mode}
\title{Upload files to public datalake with metadata}
\usage{
ellipse_push_datalake_mode(
  con,
  file_or_folder,
  dataset_name = NULL,
  tag = NULL,
  metadata = NULL,
  interactive = TRUE
)
}
\arguments{
\item{con}{Database connection to public datalake}

\item{file_or_folder}{File path or folder path to upload}

\item{dataset_name}{Dataset name for the upload}

\item{tag}{Tag name for the dataset version}

\item{metadata}{Named list of custom metadata}

\item{interactive}{Whether to use interactive mode}
}
\description{
Upload files to public datalake with metadata
}
\keyword{internal}
