% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_push_landingzone_mode}
\alias{ellipse_push_landingzone_mode}
\title{Traditional landing zone upload mode (legacy ellipse_ingest logic)}
\usage{
ellipse_push_landingzone_mode(
  con,
  file_or_folder,
  pipeline,
  file_batch = NULL,
  file_version = NULL
)
}
\description{
Traditional landing zone upload mode (legacy ellipse_ingest logic)
}
\keyword{internal}
