% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_put}
\alias{ellipse_put}
\title{Injecter un dataframe dans un datamart}
\usage{
ellipse_put(env, dataframe, datamart, table)
}
\arguments{
\item{env}{L'environnement dans lequel les données doivent être injectées}

\item{dataframe}{Le chemin vers le répertoire qui contient les fichiers à charger dans tube}

\item{datamart}{Le nom du pipeline qui doit être exécuté pour charger les données}

\item{table}{Le nom de la table qui doit être créée dans le datamart}
}
\value{
TRUE si le dataframe a été envoyé dans le datamart  FALSE sinon.
}
\description{
Injecter un dataframe dans un datamart
}
