% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_query.R
\name{ellipse_query_datalake_aggregator}
\alias{ellipse_query_datalake_aggregator}
\title{File aggregator mode for public datalake connections}
\usage{
ellipse_query_datalake_aggregator(con, dataset, tag = NULL, file = NULL)
}
\arguments{
\item{con}{Database connection to public datalake}

\item{dataset}{Dataset name to aggregate}

\item{tag}{Optional tag to filter files}

\item{file}{Optional specific filename to download/display}
}
\description{
File aggregator mode for public datalake connections
}
\keyword{internal}
