% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_query.R
\name{ellipse_query_table_mode}
\alias{ellipse_query_table_mode}
\title{Traditional table query mode for datawarehouse/datamarts}
\usage{
ellipse_query_table_mode(con, table)
}
\arguments{
\item{con}{Database connection}

\item{table}{Table name to query}
}
\description{
Traditional table query mode for datawarehouse/datamarts
}
\keyword{internal}
