% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_read}
\alias{ellipse_read}
\title{Lire et exploiter une table contenue dans l'entrepôt de données ellipse}
\usage{
ellipse_read(con, table)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{table}{Une table que l'on souhaite interroger avec \code{dplyr}.}
}
\value{
Une table Athena qui peut être interrogée dans un \emph{pipeline}
\code{dplyr}.
}
\description{
Lire et exploiter une table contenue dans l'entrepôt de données ellipse
}
