% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{extract_rows_from_elements}
\alias{extract_rows_from_elements}
\title{Extract rows from XML elements}
\usage{
extract_rows_from_elements(elements)
}
\arguments{
\item{elements}{XML nodes that represent potential rows}
}
\description{
Extract rows from XML elements
}
\keyword{internal}
