% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{extract_xml_tabular_data}
\alias{extract_xml_tabular_data}
\title{Extract tabular data from XML by finding repeating elements}
\usage{
extract_xml_tabular_data(xml_doc)
}
\arguments{
\item{xml_doc}{XML document object}
}
\description{
Extract tabular data from XML by finding repeating elements
}
\keyword{internal}
