% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{find_datalake_indexing_lambda}
\alias{find_datalake_indexing_lambda}
\title{Find the lambda function for datalake indexing}
\usage{
find_datalake_indexing_lambda(credentials)
}
\arguments{
\item{credentials}{AWS credentials from get_aws_credentials()}
}
\value{
Lambda function name if found, NULL otherwise
}
\description{
Find the lambda function for datalake indexing
}
\keyword{internal}
