% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_discover.R
\name{format_public_datalake_all_datasets}
\alias{format_public_datalake_all_datasets}
\title{Format public datalake discovery results for all datasets}
\usage{
format_public_datalake_all_datasets(con)
}
\arguments{
\item{con}{Database connection object}
}
\value{
Formatted tibble with all public datalake datasets
}
\description{
Format public datalake discovery results for all datasets
}
\keyword{internal}
