% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_public_datalake_discovery.R
\name{format_public_datalake_dataset_details}
\alias{format_public_datalake_dataset_details}
\title{Format public datalake discovery results for specific dataset}
\usage{
format_public_datalake_dataset_details(con, dataset_name)
}
\arguments{
\item{con}{Database connection object}

\item{dataset_name}{Exact dataset name}
}
\value{
List with detailed dataset information
}
\description{
Format public datalake discovery results for specific dataset
}
\keyword{internal}
