% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_discover.R
\name{format_public_datalake_pattern_search}
\alias{format_public_datalake_pattern_search}
\title{Format public datalake discovery results for pattern search}
\usage{
format_public_datalake_pattern_search(con, pattern)
}
\arguments{
\item{con}{Database connection object}

\item{pattern}{Search pattern for dataset names}
}
\value{
List with search results and summary
}
\description{
Format public datalake discovery results for pattern search
}
\keyword{internal}
