% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_discover.R
\name{format_public_datalake_tag_details}
\alias{format_public_datalake_tag_details}
\title{Format public datalake discovery results for specific dataset and tag (CONDENSED VERSION)}
\usage{
format_public_datalake_tag_details(con, dataset_name, tag_name)
}
\arguments{
\item{con}{Database connection object}

\item{dataset_name}{Exact dataset name}

\item{tag_name}{Exact tag name}
}
\value{
List with specific tag information
}
\description{
Format public datalake discovery results for specific dataset and tag (CONDENSED VERSION)
}
\keyword{internal}
