% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate test and example data}
\usage{
gen_data(model, seed = NULL, n, d, pi = 0.5, nleave = 0)
}
\arguments{
\item{model}{A character specifying the model used to generate data.
Should be one of "gaussian", "t", or "mixture".}

\item{seed}{An integer specifying the seed (optional).}

\item{n}{An integer specifying the sample size.}

\item{d}{An intger specifying the feature number/dimension.}

\item{pi}{A numeric specifying the proportion of class 0 sample. Defaults to 0.5.}

\item{nleave}{An integer specifying the size of left-out sample. For debug only.}
}
\value{
A named list of generated data. It has two elements: \code{x} and \code{y}.
}
\description{
Generate test and example data
}
