% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{get_aws_account_id}
\alias{get_aws_account_id}
\title{Get AWS Account ID from credentials}
\usage{
get_aws_account_id(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
The AWS account ID as a string
}
\description{
Get AWS Account ID from credentials
}
\keyword{internal}
