% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_column_type}
\alias{get_column_type}
\title{Déterminer le type de colonne d'in dataframe}
\usage{
get_column_type(column)
}
\arguments{
\item{column}{La colonne à analyser
Les types de colonne supportés sont: string, int, decimal et date}
}
\value{
Le type de colonne
}
\description{
Déterminer le type de colonne d'in dataframe
}
