% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{get_content_type}
\alias{get_content_type}
\title{Get appropriate content type for file}
\usage{
get_content_type(file_path)
}
\description{
Get appropriate content type for file
}
\keyword{internal}
