% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_query.R
\name{get_datalake_files_metadata}
\alias{get_datalake_files_metadata}
\title{Get file metadata from public datalake for a specific dataset}
\usage{
get_datalake_files_metadata(con, dataset, tag = NULL)
}
\arguments{
\item{con}{Database connection}

\item{dataset}{Dataset name}

\item{tag}{Optional tag filter}
}
\description{
Get file metadata from public datalake for a specific dataset
}
\keyword{internal}
