% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{get_user_choice}
\alias{get_user_choice}
\title{Handle numeric selection from the menu
Get user choice with validation}
\usage{
get_user_choice()

get_user_choice()
}
\description{
Handle numeric selection from the menu
Get user choice with validation

Get user choice with validation
}
\keyword{internal}
