% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawarehouse.R
\name{glue_table_list_to_tibble}
\alias{glue_table_list_to_tibble}
\title{Convert a list of tables from the AWS Glue API}
\usage{
glue_table_list_to_tibble(glue_response)
}
\arguments{
\item{glue_response}{A list of 2 elements, the first of which is named
\code{TableList}. What is expected here is the output from
\code{tube::list_datawarehouse_tables()}}
}
\value{
A tibble with columns:
\itemize{
\item \code{table_name} : Name of the table in the data warehouse
\item \code{col_name} : Name of the column
\item \code{col_type} : Data type of the column
\item \code{is_partition} : Logical indicating wether or not the column is
partitionned
}
}
\description{
The Glue API returns a complex \code{JSON} response when the
\code{list_database_tables} method is called. There is simply too much nested
information for a data scientist to parse through to get to the information
they need.
}
\details{
This function converts the response into a human readable \code{tibble} containing
only the most useful information, namely the table names, their respective
columns and column types, as well which columns are partitionned in the
data warehouse.
}
