% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{handle_navigation_choice}
\alias{handle_navigation_choice}
\title{Handle navigation-only choices}
\usage{
handle_navigation_choice(choice, current_dir)
}
\description{
Handle navigation-only choices
}
\keyword{internal}
