% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{handle_numeric_selection}
\alias{handle_numeric_selection}
\title{Handle numeric selection from the menu}
\usage{
handle_numeric_selection(choice_num, current_dir)
}
\description{
Handle numeric selection from the menu
}
\keyword{internal}
