% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{handle_simple_choice}
\alias{handle_simple_choice}
\title{Handle simple user choice}
\usage{
handle_simple_choice(choice, choice_map, current_dir)

handle_simple_choice(choice, choice_map, current_dir)
}
\description{
Handle simple user choice

Handle simple user choice
}
\keyword{internal}
