% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{interactive_datalake_push_flow}
\alias{interactive_datalake_push_flow}
\title{Interactive flow for collecting push parameters}
\usage{
interactive_datalake_push_flow(file_or_folder, dataset_name, tag, metadata)
}
\arguments{
\item{file_or_folder}{Initial file/folder (may be NULL)}

\item{dataset_name}{Initial dataset name (may be NULL)}

\item{tag}{Initial tag (may be NULL)}

\item{metadata}{Initial metadata (may be NULL)}
}
\description{
Interactive flow for collecting push parameters
}
\keyword{internal}
