% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_datalake.R, R/utils_datalake_push.R
\name{invoke_datalake_indexing_lambda}
\alias{invoke_datalake_indexing_lambda}
\title{Invoke AWS Lambda to index public datalake content}
\usage{
invoke_datalake_indexing_lambda(creds)

invoke_datalake_indexing_lambda(creds)
}
\arguments{
\item{creds}{AWS credentials from get_aws_credentials()}

\item{credentials}{AWS credentials from get_aws_credentials()}
}
\value{
TRUE if lambda invocation successful, FALSE otherwise
}
\description{
Trigger the lambda function to update the public-data-lake-content table
}
\keyword{internal}
