% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{is_image_dataset}
\alias{is_image_dataset}
\title{Check if files are image files}
\usage{
is_image_dataset(files)
}
\arguments{
\item{files}{Vector of file paths}
}
\value{
TRUE if all files are images, FALSE otherwise
}
\description{
Check if files are image files
}
\keyword{internal}
