% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena.R
\name{list_athena_staging_bucket}
\alias{list_athena_staging_bucket}
\title{Returns the Athena staging bucket name}
\usage{
list_athena_staging_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the Athena staging bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{athenaqueryresults}
In our infrastructure, there is only one per AWS account (DEV/PROD)
}
\details{
This is used to get the name of the bucket where the results of the
athena queries are stored.  More specifically, it is used to create
the \code{s3_staging_dir} parameter in the \code{DBI::dbConnect} function call
in ellipse_connect
}
