% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalake.R
\name{list_datalake_bucket}
\alias{list_datalake_bucket}
\title{Returns the datalake bucket name}
\usage{
list_datalake_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the datalake bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{datalakebucket}
}
\details{
In our infrastructure, there is only one per AWS account (DEV/PROD)
}
