% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamart.R
\name{list_datamarts_bucket}
\alias{list_datamarts_bucket}
\title{Returns the datamarts bucket name}
\usage{
list_datamarts_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the datamart bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{datamartbucket}
}
\details{
In our infrastructure, there is only one per AWS account (DEV/PROD)

It is used mainly by the ellipse_publish function to store R dataframes
in a datamart in the form of CSV files, which in turn, are converted into
GLUE tables by the glue job.
}
