% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawarehouse.R
\name{list_datawarehouse_bucket}
\alias{list_datawarehouse_bucket}
\title{Returns the datawarehouse bucket name}
\usage{
list_datawarehouse_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the datamart bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{datawarehousebucket} but in our
infrastructure, there is only one per AWS account (DEV/PROD)
}
\details{
Currently the function is not used in the package but it is
kept for future use
}
