% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datawarehouse.R
\name{list_datawarehouse_database}
\alias{list_datawarehouse_database}
\title{Returns the datawarehouse GLUE database name}
\usage{
list_datawarehouse_database(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the datawarehouse database
}
\description{
Technically the function returns all the databases of which
the name contains the string \code{datamart} but in our
infrastructure, there is only one per AWS account (DEV/PROD)
}
\details{
It is used mainly by the ellipse_connect function to retrieve the
schema with wich to instanciate the DBI connection
}
