% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{list_glue_databases}
\alias{list_glue_databases}
\title{List all the databases in the AWS Glue Data Catalog}
\usage{
list_glue_databases(credentials, type)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{type}{A string to filter the databases by (e.g. "datawarehouse" or "datamart")}
}
\value{
A list of databases
}
\description{
List all the databases in the AWS Glue Data Catalog
}
