% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{list_glue_jobs}
\alias{list_glue_jobs}
\title{Lists all the GLUE jobs in the account}
\usage{
list_glue_jobs(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A list of jobs
}
\description{
Lists all the GLUE jobs in the account
}
