% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{list_glue_table_properties}
\alias{list_glue_table_properties}
\title{List the properties of a GLUE table}
\usage{
list_glue_table_properties(credentials, schema, table)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{schema}{The DBI schema to list the tables from}

\item{table}{The name of the table to list the properties from}
}
\value{
A tibble with the properties of the table
}
\description{
List the properties of a GLUE table
}
\examples{
list_glue_table_properties(credentials, DBI::dbGetInfo(con)$dbms.name, "my_table")
}
