% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{list_glue_tables}
\alias{list_glue_tables}
\title{Lists all the tables in a GLUE database}
\usage{
list_glue_tables(credentials, schema, tablename_filter = NULL, simplify = TRUE)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{schema}{The DBI schema to list the tables from}

\item{tablename_filter}{A string to filter the tables by}

\item{simplify}{A boolean indicating wether or not to simplify the output}
}
\value{
A list of tables
}
\description{
Lists all the tables in a GLUE database
}
