% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{list_lambda_functions}
\alias{list_lambda_functions}
\title{List all Lambda functions with robust AWS API pagination}
\usage{
list_lambda_functions(credentials)
}
\arguments{
\item{credentials}{AWS credentials from get_aws_credentials()}
}
\value{
Vector of lambda function names
}
\description{
List all Lambda functions with robust AWS API pagination
}
\keyword{internal}
