% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{list_landing_zone_bucket}
\alias{list_landing_zone_bucket}
\title{Returns the landing zone bucket name}
\usage{
list_landing_zone_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the datalake bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{datalakebucket} but in our infrastructure,
there is only one per AWS account (DEV/PROD)
}
\details{
The landing zone is used by ellipse_ingest to allow users
to ingest raw data into the dataplatform so that it gets
processed by the pipeline (extractor - loader)
}
