% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_datalake.R
\name{list_public_datalake_bucket}
\alias{list_public_datalake_bucket}
\title{Returns the public datalake bucket name}
\usage{
list_public_datalake_bucket(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the public datalake bucket
}
\description{
Technically the function returns all the buckets for which
the name contains the string \code{publicdatalakebucket}
}
\details{
In our infrastructure, there is only one per AWS account (DEV/PROD)

It is used mainly by the ellipse_connect function and related
discovery functions to interact with user-uploaded files
in the public datalake.
}
