% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public-datalake.R
\name{list_public_datalake_database}
\alias{list_public_datalake_database}
\title{Returns the public datalake GLUE database name}
\usage{
list_public_datalake_database(credentials)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}
}
\value{
A string containing the name of the public datalake database
}
\description{
Technically the function returns all the databases of which
the name contains the string \code{publicdatalake} but in our
infrastructure, there is only one per AWS account (DEV/PROD)
}
\details{
It is used mainly by the ellipse_connect function to retrieve the
schema with which to instantiate the DBI connection for the
public datalake containing user-uploaded files with metadata.
}
