% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{list_s3_folders}
\alias{list_s3_folders}
\title{List S3 folders}
\usage{
list_s3_folders(credentials, bucket, prefix)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{bucket}{The bucket to list folders from}

\item{prefix}{The prefix (or partition) to list folders from}
}
\value{
A character vector of S3 folders
}
\description{
List all the S3 folders within a particular prefix (or partition) in a bucket
}
