% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{list_s3_partitions}
\alias{list_s3_partitions}
\title{List S3 partitions}
\usage{
list_s3_partitions(credentials, bucket)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package
#' @param bucket The bucket to list partitions from}
}
\value{
A character vector of S3 partitions
}
\description{
List all the S3 partitions in a bucket even the empty ones (containing no object)
}
