% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{merge_dataframes_with_schema_union}
\alias{merge_dataframes_with_schema_union}
\title{Merge multiple dataframes with schema union}
\usage{
merge_dataframes_with_schema_union(df_list)
}
\arguments{
\item{df_list}{List of dataframes to merge}
}
\description{
Merge multiple dataframes with schema union
}
\keyword{internal}
