% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{prepare_files_for_upload}
\alias{prepare_files_for_upload}
\title{Prepare files for upload (handle both files and folders)}
\usage{
prepare_files_for_upload(file_or_folder)
}
\description{
Prepare files for upload (handle both files and folders)
}
\keyword{internal}
