% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{prepare_s3_metadata}
\alias{prepare_s3_metadata}
\title{Prepare metadata for S3 upload}
\usage{
prepare_s3_metadata(custom_metadata)
}
\description{
Prepare metadata for S3 upload
}
\keyword{internal}
