% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-tools.R
\name{read_dat_with_overflow_handling}
\alias{read_dat_with_overflow_handling}
\title{Read DAT file with overflow handling}
\usage{
read_dat_with_overflow_handling(filepath)
}
\arguments{
\item{filepath}{Path to DAT file}
}
\description{
Read DAT file with overflow handling
}
\keyword{internal}
