% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{read_file_by_extension}
\alias{read_file_by_extension}
\title{Read file based on extension with overflow handling}
\usage{
read_file_by_extension(filepath, extension)
}
\arguments{
\item{filepath}{Path to file}

\item{extension}{File extension}
}
\description{
Read file based on extension with overflow handling
}
\keyword{internal}
