% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{read_rda_file}
\alias{read_rda_file}
\title{Read RDA file (load and return first object)}
\usage{
read_rda_file(filepath)
}
\arguments{
\item{filepath}{Path to RDA file}
}
\description{
Read RDA file (load and return first object)
}
\keyword{internal}
