% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{read_xml_as_tabular}
\alias{read_xml_as_tabular}
\title{Read XML file and convert to tabular format intelligently}
\usage{
read_xml_as_tabular(filepath)
}
\arguments{
\item{filepath}{Path to XML file}
}
\description{
Read XML file and convert to tabular format intelligently
}
\keyword{internal}
