% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{run_glue_job}
\alias{run_glue_job}
\title{Run a GLUE job manually specifically for ingesting CSV in to
the datawarehouse or datamart}
\usage{
run_glue_job(
  credentials,
  job_name,
  database,
  prefix,
  table_tags = NULL,
  table_description = NULL
)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{job_name}{The name of the job to run}

\item{database}{The name of the database to run the job for
it must be either "datamart", "datamarts" or "datawarehouse"}

\item{prefix}{The prefix to run the job for.  This is the first
level partition in the datawarehouse or datamart bucket
in the datawarehouse, it represents the pipeline name
in the datamart, it represents the datamart and table name separates with a /}

\item{table_tags}{Optional tags to associate with the table}

\item{table_description}{Optional description for the table}
}
\value{
A boolean indicating wether or not the job was started
}
\description{
Run a GLUE job manually specifically for ingesting CSV in to
the datawarehouse or datamart
}
