% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{simple_file_folder_selector}
\alias{simple_file_folder_selector}
\title{Simple file and folder selector with numbered choices}
\usage{
simple_file_folder_selector()
}
\description{
Simple file and folder selector with numbered choices
}
\keyword{internal}
