% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{stratification}
\alias{stratification}
\title{Stratify training data}
\usage{
stratification(xtrain, ytrain, wcost)
}
\arguments{
\item{xtrain}{A matrix with columns representing features and rows representing observations.}

\item{ytrain}{A binary vector specifying the corresponding labels of the observations.}

\item{wcost}{A numeric specifying the type I error cost used for stratification.}
}
\value{
A named list specifying the stratified data. It has two elements: \code{x} and \code{y}.
}
\description{
Stratify training data
}
