% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tube.R
\name{TUBE}
\alias{TUBE}
\title{TUBE-assisted cost-sensitive classification}
\usage{
TUBE(
  data,
  classify_fun,
  alpha = 0.05,
  delta = 0.1,
  t_cs = 0.5,
  nleave = NULL,
  cost = seq(0.1, 0.99, 0.01),
  B1 = 50,
  B2 = 200
)
}
\arguments{
\item{data}{A named list specifying the training data. It should have two elements:
\code{data$x} should be a matrix with columns representing features and rows representing observations.
\code{data$y} should be a binary vector specifying the corresponding labels of the observations.}

\item{classify_fun}{A function specifying the cost-sensitive classification algorithm.}

\item{alpha}{A numeric specifying the target type I error (between 0 and 1). Defaults to 0.05.}

\item{delta}{A numeric specifying the target violation rate (between 0 and 1). Defaults to 0.1.}

\item{t_cs}{A numeric specifying the threshold applied to classification scores. Defaults to 0.5.}

\item{nleave}{A integer specifying the sample size of left-out class 0 data. Should be smaller than the
number of class 0 cases in \code{data}.}

\item{cost}{A numeric vector specifying the candidate type I error costs. Should be ordered increasingly.
Defaults to \code{seq(0.1, 0.99, 0.01)}.}

\item{B1}{An integer specifying the number of random data splitting. Defaults to 50.}

\item{B2}{An integer specifying the number of bootstrap samples. Defaults to 200.}
}
\value{
A list with two elements. \code{c0} gives the selected type I error cost;
\code{yhat} gives the predicted class labels for \code{data$x} given \code{c0}.
}
\description{
TUBE-assisted cost-sensitive classification
}
\references{
Li WV, Tong X, Li JJ. Bridging Cost-sensitive and Neyman-Pearson Paradigms for Asymmetric Binary Classification.
arXiv preprint arXiv:2012.14951. 2020 Dec 29. \url{https://arxiv.org/abs/2012.14951}
}
\author{
Wei Vivian Li, \email{vivian.li@rutgers.edu}

Xin Tong, \email{xtong001@gmail.com>}

Jingyi Jessica Li, \email{jli@stat.ucla.edu}
}
