% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{upload_file_to_s3}
\alias{upload_file_to_s3}
\title{Upload a file to an S3 bucket}
\usage{
upload_file_to_s3(credentials, file, bucket, key)
}
\arguments{
\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{file}{The file to upload}

\item{bucket}{The bucket to upload the file to}

\item{key}{The key to use for the file}
}
\value{
TRUE if the file was uploaded successfully
}
\description{
Upload a file to an S3 bucket
}
