% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{upload_files_to_public_datalake}
\alias{upload_files_to_public_datalake}
\title{Upload files to public datalake S3 bucket}
\usage{
upload_files_to_public_datalake(creds, files, dataset_name, tag, metadata)
}
\description{
Upload files to public datalake S3 bucket
}
\keyword{internal}
