% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{validate_datalake_push_params}
\alias{validate_datalake_push_params}
\title{Validate parameters for datalake push}
\usage{
validate_datalake_push_params(file_or_folder, dataset_name, tag, metadata)
}
\description{
Validate parameters for datalake push
}
\keyword{internal}
