% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_dbi.R
\name{athena_coerce_column}
\alias{athena_coerce_column}
\title{Convertir les lignes brutes retournées par Athena en un vecteur R typé}
\usage{
athena_coerce_column(x, type)
}
\arguments{
\item{x}{Un vecteur de chaînes de caractères}

\item{type}{Le type de colonne Athena (ex.: "varchar", "bigint", "double")}
}
\value{
Un vecteur R du type approprié
}
\description{
Convertir les lignes brutes retournées par Athena en un vecteur R typé
}
\keyword{internal}
