% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check_params.R
\name{check_params_before_publish}
\alias{check_params_before_publish}
\title{Check the parameters provided to the ellipse_publish function}
\usage{
check_params_before_publish(env, dataframe, datamart, table, tag)
}
\arguments{
\item{env}{The environment to publish the data to}

\item{dataframe}{The dataframe to publish}

\item{table}{The table to publish the data to}
}
\value{
TRUE if the parameters are valid, FALSE otherwise
}
\description{
Check if the parameters are valid before publishing the data
}
