% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_datalake_push.R
\name{collect_required_system_metadata}
\alias{collect_required_system_metadata}
\title{Collect required system metadata fields}
\usage{
collect_required_system_metadata(metadata)
}
\description{
Collect required system metadata fields
}
\keyword{internal}
