% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_url.R
\name{convert_url_to_key}
\alias{convert_url_to_key}
\title{convert a url into a string}
\usage{
convert_url_to_key(url)
}
\arguments{
\item{url}{The URL to convert to a key string}
}
\value{
A cleaned string suitable for use as a key
}
\description{
convert a url into a string
}
\examples{
\dontrun{
r <- convert_url_to_key(url)
print(r)
}

}
