% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_connect}
\alias{ellipse_connect}
\title{Se connecter à la plateforme de données ellipse sur AWS}
\usage{
ellipse_connect(env = c("PROD", "DEV"), database = "datawarehouse")
}
\arguments{
\item{env}{The environment to connect to on ellipse-science. Defaults to "PROD"}

\item{database}{The Glue/Athena database to connect to. Default to "datawarehouse"}
}
\value{
Un object de connexion \code{DBI}.
}
\description{
Cette fonction utilise les clés d'accès AWS configurées dans le fichier
\code{.Renviron} pour se connecter à la plateforme de données.
}
