% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_discover}
\alias{ellipse_discover}
\title{Découvrir les tables disponibles sur la plateforme ellipse, ainsi que leur
contenu}
\usage{
ellipse_discover(con, table = NULL)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{table}{Une table pour laquelle on veut obtenir les informations.}
}
\value{
Un \code{tibble} contenant les tables diposnibles dans l'entrepôt de
données, ou un descriptino des colonnes pour une table en particulier.
}
\description{
Si aucune table n'est passée en paramètre, un sommaire des tables disponibles
dans l'entrepôt de données est retourné. Si un nom de \code{table} est passé en
paramètre, une description des colonnes de cette table est retournée.
}
