% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_ingest}
\alias{ellipse_ingest}
\title{Injecter de nouvelles données brutes manuellement dans tube via la landing zone}
\usage{
ellipse_ingest(
  con,
  file_or_folder,
  pipeline,
  file_batch = NULL,
  file_version = NULL
)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{file_or_folder}{Le chemin vers le répertoire qui contient les fichiers à charger dans tube}

\item{pipeline}{Le nom du pipeline qui doit être exécuté pour charger les données.}

\item{file_batch}{Le nom du batch qui doit être accollé aux données dans l'entrepôt de
données. Utilisé pour les données factuelles seulement, NULL sinon.}

\item{file_version}{La version des données qui doit être accollée aux données dans
l'entrepôt de données. Utilisé pour les données dimensionnelles et les dictionnaires
seulement, NULL sinon.}
}
\value{
La liste des fichiers qui ont été injectés dans tube
}
\description{
DEPRECATED: Utilisez \code{ellipse_push()} à la place.
Cette fonction est maintenue pour la compatibilité descendante.
}
