% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_ingest}
\alias{ellipse_ingest}
\title{Injecter de nouvelles données brutes manuellement dans tube via la landing zone}
\usage{
ellipse_ingest(env, folder, pipeline, batch)
}
\arguments{
\item{env}{L'environnement dans lequel les données doivent être injectées}

\item{folder}{Le chemin vers le répertoire qui contient les fichiers à charger dans tube}

\item{pipeline}{Le nom du pipeline qui doit être exécuté pour charger les données}

\item{batch}{Le nom du batch qui doit être accollé aux données dans l'entrepôt de données}
}
\value{
La liste des fichiers qui ont été injectés dans tube
}
\description{
Injecter de nouvelles données brutes manuellement dans tube via la landing zone
}
