% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_partitions}
\alias{ellipse_partitions}
\title{Obtenir le domaine de valeurs pour les dimensions d'une table}
\usage{
ellipse_partitions(con, table)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{table}{Une table pour laquelle on veut obtenir les informations.}
}
\value{
Un \code{tibble} contenant le nombre d'observations par valeur de
groupement des variables partitionnées.
}
\description{
Obtenir le domaine de valeurs pour les dimensions d'une table
}
\details{
Les tables de données sur AWS sont segmentées en \emph{partitions}. Les requêtes
qui ciblent une plage précise dans ces partitions réduisent les coûts
d'utilisation de la plateforme, parce que les données à l'extérieur de cette
plage ne sont pas lues par \emph{AWS Athena}.

Cette fonction permet d'obtenir les valeurs possibles des partitions d'une
table donnée afin de mieux cibler nos requêtes ensuite.
}
