% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_publish}
\alias{ellipse_publish}
\title{Publier un dataframe dans un datamart}
\usage{
ellipse_publish(con, dataframe, datamart, table, tag = NULL)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{dataframe}{Le chemin vers le répertoire qui contient les fichiers à charger dans tube}

\item{datamart}{Le nom du pipeline qui doit être exécuté pour charger les données}

\item{table}{Le nom de la table qui doit être créée dans le datamart}
}
\value{
TRUE si le dataframe a été envoyé dans le datamart  FALSE sinon.
}
\description{
Publier un dataframe dans un datamart
}
