% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_query}
\alias{ellipse_query}
\title{Lire et exploiter une table contenue dans l'entrepôt de données ellipse}
\usage{
ellipse_query(con, dataset, tag = NULL)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{dataset}{Pour les connexions datawarehouse/datamarts: nom d'une table.
Pour les connexions datalake: nom d'un dataset à agréger.}

\item{tag}{Optionnel. Pour les connexions datalake seulement: tag spécifique
à filtrer. Si NULL, agrège tous les tags du dataset.}
}
\value{
Pour datawarehouse/datamarts: Une table Athena qui peut être interrogée
dans un pipeline \code{dplyr}. Pour datalake datasets: Un dataframe agrégé de tous les
fichiers du dataset. Pour datalake images: Interface interactive pour visualiser
les images une par une.
}
\description{
Lire et exploiter une table contenue dans l'entrepôt de données ellipse
}
