% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse_unpublish}
\alias{ellipse_unpublish}
\title{Retirer une table d'un datamart ou un datamart complet}
\usage{
ellipse_unpublish(con, datamart, table)
}
\arguments{
\item{con}{Un objet de connexion tel qu'obtenu via \code{tube::ellipse_connect()}.}

\item{datamart}{Le nom du datamart contenant la table à retirer.}

\item{table}{Paramètre optionnel : Le nom de la table à retirer.  S'il est manquant,
vide ou null alors c'est le datamart complet qui est retiré}
}
\value{
TRUE si la table a été retirée avec succès, FALSE sinon.
}
\description{
Retirer une table d'un datamart ou un datamart complet
}
