% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand side}

\item{y}{Right-hand side}
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns the right-hand side.
}
\keyword{internal}
