% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{update_glue_table_tags}
\alias{update_glue_table_tags}
\title{Update the custom tags (advanced properties) of a glue table}
\usage{
update_glue_table_tags(creds, schema, table, new_table_tags)
}
\arguments{
\item{schema}{The DBI schema to list the tables from}

\item{table}{The name of the table to update the tags of}

\item{credentials}{A list of AWS credentials in the format compliant
with the paws package}

\item{table_tags}{A list of custom tags to update the table with}
}
\value{
A boolean indicating wether or not the tags were updated
successfully
}
\description{
Update the custom tags (advanced properties) of a glue table
}
\examples{
update_glue_table_tags(credentials,
                       DBI::dbGetInfo(con)$dbms.name,
                       "my_table",
                       list(tag1 = "value1", tag2 = "value2"))
}
