% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landing.R
\name{upload_to_landing_zone}
\alias{upload_to_landing_zone}
\title{Uploads the files in the specified local folder to the landing zone bucket.  The landing zone takes a maximum of 30 files
at a time.  If your local_folder contains more than 30 files, the files will be segmented into sets of 30 files and uploaded
one set at a time.  The function will wait 6 minutes between each set to ensure that the data platform has time to process the files.
Please ensure that your pipeline has been configured by your data platform administrator before uploading data to the landing zone, otherwise
the function will return an error message.}
\usage{
upload_to_landing_zone(
  creds,
  local_folder,
  pipeline_name,
  batch = NULL,
  timestamp_files = FALSE
)
}
\arguments{
\item{creds}{An AWS session object with your credentials and the aws ressources required}

\item{local_folder}{The path to the local folder containing the files to upload}

\item{pipeline_name}{The name of the pipeline (i.e.: the name of the first folder in the path within the landding zone bucket)}

\item{batch}{mandatory parameter to specify the batch name.  If not provided, the date and time batch will be used}

\item{timestamp_files}{If TRUE, the files will be timestamped with the current date and time.
This will ensure that the files are unique in the landing zone bucket}
}
\value{
the status of each file upload
}
\description{
Uploads the files in the specified local folder to the landing zone bucket.  The landing zone takes a maximum of 30 files
at a time.  If your local_folder contains more than 30 files, the files will be segmented into sets of 30 files and uploaded
one set at a time.  The function will wait 6 minutes between each set to ensure that the data platform has time to process the files.
Please ensure that your pipeline has been configured by your data platform administrator before uploading data to the landing zone, otherwise
the function will return an error message.
}
\examples{
\dontrun{
 r <- upload_to_landing_zone(aws_session(), "my_filder", "my_pipeline", NULL, TRUE)
print(r)
}
}
